requirements.txt
pip==24.0
trulens-core @ git+https://github.com/truera/trulens#egg=trulens-core&subdirectory=src/core/
trulens-feedback @ git+https://github.com/truera/trulens#egg=trulens-feedback&subdirectory=src/feedback/
trulens-providers-cortex @ git+https://github.com/truera/trulens#egg=trulens-providers-cortex&subdirectory=src/providers/cortex/
trulens-dashboard @ git+https://github.com/truera/trulens#egg=trulens-dashboard&subdirectory=src/dashboard/
snowflake-sqlalchemy
snowflake-snowpark-python
snowflake-ml-python
snowflake.core
